Remove D3.js dependency from rendering svg elements
