+++ tests/gui/view/pca.test.js
@@ -1,0 +1,44 @@
+import { getPage } from '../helper/browser'
+
+describe('dimensionality reduction', () => {
+	/** @type {Awaited<ReturnType<getPage>>} */
+	let page
+	beforeEach(async () => {
+		page = await getPage()
+	}, 10000)
+
+	afterEach(async () => {
+		await page?.close()
+	})
+
+	test('initialize', async () => {
+		const taskSelectBox = await page.waitForSelector('#ml_selector dl:first-child dd:nth-child(5) select')
+		await taskSelectBox.selectOption('DR')
+		const modelSelectBox = await page.waitForSelector('#ml_selector .model_selection #mlDisp')
+		await modelSelectBox.selectOption('pca')
+		const methodMenu = await page.waitForSelector('#ml_selector #method_menu')
+		const buttons = await methodMenu.waitForSelector('.buttons')
+
+		const methods = await buttons.waitForSelector('select:nth-of-type(1)')
+		await expect((await methods.getProperty('value')).jsonValue()).resolves.toBe('')
+		const fitButton = await buttons.waitForSelector('input[value=Fit]')
+		expect(fitButton).toBeDefined()
+	}, 10000)
+
+	test('learn', async () => {
+		const taskSelectBox = await page.waitForSelector('#ml_selector dl:first-child dd:nth-child(5) select')
+		await taskSelectBox.selectOption('DR')
+		const modelSelectBox = await page.waitForSelector('#ml_selector .model_selection #mlDisp')
+		await modelSelectBox.selectOption('pca')
+		const methodMenu = await page.waitForSelector('#ml_selector #method_menu')
+		const buttons = await methodMenu.waitForSelector('.buttons')
+
+		const fitButton = await buttons.waitForSelector('input[value=Fit]')
+		await fitButton.evaluate(el => el.click())
+
+		const svg = await page.waitForSelector('#plot-area svg')
+		await svg.waitForSelector('.tile circle')
+		const circles = await svg.$$('.tile circle')
+		expect(circles).toHaveLength(300)
+	}, 60000)
+})
